requirements.txt
flask-cors==6.0.2
g4f==7.5.0
gunicorn==23.0.0
openpyxl==3.1.5
pdfplumber==0.11.9
python-docx==1.2.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
